ETag depends on Accept-Encoding but not on URL
Removed the dependence of ETag on URL. Instead the computed ETag depends
on the presence and value of the Accept-Encoding header in the request.
This is an easy to implement approximation of the ideal case, where ETag
should depend only on the actual resolved Content-Encoding of the
response. Such a solution will work well with browser caches but may
confuse proxies.